// Front/src/app/model/General.ts
export interface General {
    codigo?: string;
    clave?: string;
    descripcion1?: string;
    descripcionPrimero?: string;
    general?:string;
    estado?: boolean;
    usuarioCreacion?: string;
    fechaCreacion?:string;
    horaCreacion?: string;
    usuarioActualizacion?: string;
    fechaActualizacion?: Date;
    horaActualizacion?: string;
  }
  